Oh Boy
Objects can be loaded
Movement
Diffuse Shading
Render Systems and renderer
Camera
Buffers
Descriptors (WIP)
balls

+++ src/graphics/swap_chain.cpp
@@ -31,7 +31,7 @@ nova_SwapChain::~nova_SwapChain() {
 		swapChain = nullptr;
 	}
 
-	for (int i = 0; i < depthImages.size(); i++) {
+	for (size_t i = 0; i < depthImages.size(); i++) {
 		vkDestroyImageView(device.device(), depthImageViews[i], nullptr);
 		vkDestroyImage(device.device(), depthImages[i], nullptr);
 		vkFreeMemory(device.device(), depthImageMemorys[i], nullptr);
@@ -275,7 +275,7 @@ void nova_SwapChain::createDepthResources() {
 	depthImageMemorys.resize(imageCount());
 	depthImageViews.resize(imageCount());
 
-	for (int i = 0; i < depthImages.size(); i++) {
+	for (size_t i = 0; i < depthImages.size(); i++) {
 		VkImageCreateInfo imageInfo{};
 		imageInfo.sType = VK_STRUCTURE_TYPE_IMAGE_CREATE_INFO;
 		imageInfo.imageType = VK_IMAGE_TYPE_2D;
